requirements.txt
flask 
flask-MySQL
flask-SQLAlchemy
flask-Migrate
flask-wtf
mysql-connector-python
pymysql
cryptography